% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbrd.R
\name{calculate_inbreeding}
\alias{calculate_inbreeding}
\title{Calculate population average inbreeding from effective population size}
\usage{
calculate_inbreeding(Nm, Nf)
}
\arguments{
\item{Nm}{A numeric vector of number of male parents contributing offspring in each generation}

\item{Nf}{A numeric vector of number of female parents contributing offspring in each generation}
}
\value{
A data frame
}
\description{
Calculate population average inbreeding from effective population size
}
\examples{
Nm <- c(26,40,32,32,58,54,54,64,16)  # Number of male parents in each generation
Nf <- c(16,40,40,40,64,54,54,54,18)  # Number of female parents in each generation
result <- calculate_inbreeding(Nm, Nf)
print(result)
}
